# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bunkarr API",
    "version": "1",
    "description": "Bunkarr REST API. Authenticate with the API key from Settings > General in the X-Api-Key header (or the apikey query parameter), or with a login session cookie. Errors are {\"message\": \"...\"}: 400 invalid input, 401 not authenticated, 403 refused (a cross-site browser request, an action that needs a login session, a directory Bunkarr cannot read), 404 unknown id, 409 conflict with the current state (including safety refusals that need a confirmation or a mounted share), 502 Plex failed, 503 unavailable (the backup routes of a server without backup services, or a request stopped as the server shuts down), 500 anything else. Secrets (Plex tokens, Apprise URLs) are write-only. Job-starting endpoints answer 202 with the queued Job. Sizes are bytes, times RFC 3339.",
    "license": {
      "name": "GPL-3.0-or-later",
      "identifier": "GPL-3.0-or-later"
    }
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "apiKeyHeader": []
    },
    {
      "apiKeyQuery": []
    },
    {
      "session": []
    }
  ],
  "components": {
    "securitySchemes": {
      "apiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      },
      "apiKeyQuery": {
        "type": "apiKey",
        "in": "query",
        "name": "apikey"
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "bunkarr_session"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "JobType": {
        "type": "string",
        "enum": [
          "scan",
          "sync",
          "plexdb_backup",
          "retention",
          "verify"
        ]
      },
...
      "Job": {
        "type": "object",
        "required": [
          "id",
          "type",
          "status",
          "trigger",
          "dryRun",
          "params",
          "attempt",
          "progress",
          "stats",
          "warnings",
          "summary",
          "queuedAt",
          "startedAt",
          "finishedAt"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "type": {
            "$ref": "#/components/schemas/JobType"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "trigger": {
            "$ref": "#/components/schemas/JobTrigger"
          },
          "dryRun": {
            "type": "boolean"
          },
          "params": {
            "$ref": "#/components/schemas/JobParams"
          },
          "attempt": {
            "type": "integer",
            "minimum": 1
          },
          "progress": {
            "$ref": "#/components/schemas/Progress"
          },
          "stats": {
            "type": [
              "object",
              "null"
            ],
            "description": "Runner-specific counters, e.g. sync: filesPlanned, filesCopied, filesUpdated, filesMoved, filesAdopted, filesLinked, filesRetained, filesDisplaced, filesHeld, filesFailed, filesSkipped, bytesPlanned, bytesCopied, durationMs."
          },
          "warnings": {
            "type": "integer"
          },
          "summary": {
            "type": "string"
          },
          "error": {
            "type": "string",
            "description": "Why the job failed (secrets redacted)."
          },
          "queuedAt": {
            "type": "string",
            "format": "date-time"
          },
          "startedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "finishedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
...
      "IntegrationType": {
        "type": "string",
        "enum": [
          "plex",
          "sonarr",
          "radarr",
          "lidarr",
          "tautulli",
          "seerr",
          "maintainerr"
        ]
      },
...
      "Integration": {
        "type": "object",
        "required": [
          "id",
          "type",
          "name",
          "url",
          "enabled",
          "hasApiKey",
          "settings",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "type": {
            "$ref": "#/components/schemas/IntegrationType"
          },
          "name": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "hasApiKey": {
            "type": "boolean",
            "description": "The API key (Plex token) itself is never returned."
          },
          "settings": {
            "type": "object",
            "description": "PlexSettings for Plex integrations."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "IntegrationInput": {
        "type": "object",
        "description": "type is required on create and cannot change.",
        "required": [
          "name",
          "url"
        ],
        "properties": {
          "type": {
            "$ref": "#/components/schemas/IntegrationType"
          },
          "name": {
            "type": "string",
            "maxLength": 64
          },
          "url": {
            "type": "string",
            "description": "http(s) base URL without credentials, query or fragment"
          },
          "enabled": {
            "type": "boolean"
          },
          "apiKey": {
            "type": "string",
            "description": "Write-only. Empty or missing on update keeps the stored key."
          },
          "clearApiKey": {
            "type": "boolean",
            "description": "Update only: remove the stored key."
          },
          "settings": {
            "type": "object",
            "description": "PlexSettings for Plex integrations; missing on update keeps them."
          }
        }
      },
      "IntegrationTest": {
        "type": "object",
        "required": [
          "url"
        ],
        "anyOf": [
          {
            "required": [
              "type"
            ]
          },
          {
            "required": [
              "id"
            ],
            "properties": {
              "id": {
                "minimum": 1
              }
            }
          }
        ],
        "properties": {
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/IntegrationType"
              }
            ],
            "description": "Required unless id is given (then it defaults to the saved integration's type)"
          },
          "url": {
            "type": "string"
          },
          "apiKey": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "description": "A saved integration whose stored key is used when apiKey is empty (0: none)",
            "minimum": 0
          }
        }
      },
      "IntegrationTestResult": {
        "type": "object",
        "required": [
          "ok",
          "message"
        ],
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "machineIdentifier": {
            "type": "string"
          },
          "butlerStartHour": {
            "type": "integer",
            "minimum": 0,
            "maximum": 23
          },
          "butlerEndHour": {
            "type": "integer",
            "minimum": 0,
            "maximum": 23
          }
        }
      },
...
      "PlexSection": {
        "type": "object",
        "required": [
          "key",
          "title",
          "type",
          "locations"
        ],
        "properties": {
          "key": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "locations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlexLocation"
            }
          }
        }
      },
      "PlexBackupRequest": {
        "type": "object",
        "properties": {
          "destinationId": {
            "type": "integer",
            "description": "0 or missing: the integration's backup destination"
          },
          "dryRun": {
            "type": "boolean"
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid input",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid credentials",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "No such object",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Conflict": {
        "description": "Conflicts with the current state (the message says why)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ServerError": {
        "description": "Internal error (message redacted of secrets)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "BadGateway": {
        "description": "Plex answered badly or not at all",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "No backup services are configured on this server, or the request was stopped (the server is shutting down)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "parameters": {
      "id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer",
          "minimum": 1
        }
      },
...
    }
  },
  "paths": {
...
    "/integrations": {
      "get": {
        "operationId": "listIntegrations",
        "summary": "List integrations",
        "responses": {
          "200": {
            "description": "Integrations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Integration"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "tags": [
          "Integrations"
        ]
      },
      "post": {
        "operationId": "createIntegration",
        "summary": "Add an integration (a Plex backup schedule is created from settings.backup)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntegrationInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Integration"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "tags": [
          "Integrations"
        ]
      }
    },
    "/integrations/test": {
      "post": {
        "operationId": "testIntegration",
        "summary": "Test a URL and API key (Plex: identity, then the token)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntegrationTest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Result (a failed connection is ok: false)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationTestResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "tags": [
          "Integrations"
        ]
      }
    },
    "/integrations/{id}": {
      "get": {
        "operationId": "getIntegration",
        "summary": "Get an integration",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Integration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Integration"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "tags": [
          "Integrations"
        ]
      },
      "put": {
        "operationId": "updateIntegration",
        "summary": "Update an integration; an empty apiKey keeps the stored one",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntegrationInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Integration"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "tags": [
          "Integrations"
        ],
        "description": "Changing the URL requires the token again (apiKey) or clearApiKey: a stored token is only sent to the URL it was saved with; otherwise 400."
      },
      "delete": {
        "operationId": "deleteIntegration",
        "summary": "Delete an integration and its schedules (backups are kept)",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "204": {
            "description": "Done"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "A job of the integration (a Plex DB backup) is queued or running",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "tags": [
          "Integrations"
        ]
      }
    },
    "/integrations/{id}/plex/sections": {
      "get": {
        "operationId": "listPlexSections",
        "summary": "Plex library sections with their folders mapped to Bunkarr's paths",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Sections",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlexSection"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "tags": [
          "Integrations"
        ]
      }
    },
    "/integrations/{id}/plex/backup": {
      "post": {
        "operationId": "startPlexBackup",
        "summary": "Back up the Plex database now",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlexBackupRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "The queued job (an identical queued job is returned instead of a new one)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "tags": [
          "Integrations"
        ]
      }
    },
...
    "/schedules": {
      "get": {
        "operationId": "listSchedules",
        "summary": "List schedules with their next run",
        "responses": {
          "200": {
            "description": "Schedules",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Schedule"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "tags": [
          "Jobs"
        ]
      }
    },
...
  }
}